+++ LMS.Admin.Web/wwwroot/js/socket/UserHub.js
@@ -1,5 +1,5 @@
 ﻿var connection = new signalR.HubConnectionBuilder()
-    .withUrl("/testHub")
+    .withUrl("/sessionHub")
     .configureLogging(signalR.LogLevel.Information)
     .build();
